requirements.txt
# Core Parsing
dnspython==2.4.2
beautifulsoup4==4.12.2
tldextract==5.1.1
python-dotenv==1.0.0

# Terminal UI & Export
rich==13.7.0
reportlab==4.0.8

# Intelligence Integrations
requests==2.31.0

# Malware and Attachment Analysis
# yara-python==3.11.0
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
oletools==0.60.2
pdfid==1.1.3
